+++ src/config/startupTerminal.ts
@@ -2,10 +2,10 @@ import { IsString, IsNotEmpty, IsOptional } from 'class-validator';
 
 export default class StartupTerminal {
   public readonly id: string;
-  public readonly startupCommand?: string;
+  public readonly startupCommands: string[];
 
-  constructor(id: string, startupCommand?: string) {
+  constructor(id: string, startupCommands: string[] = []) {
     this.id = id;
-    this.startupCommand = startupCommand;
+    this.startupCommands = startupCommands;
   }
 }